[Refactor] Add root exports to /Styleguide
