+++ src/Apps/Artist/Routes/AuctionResults/ArtistAuctionResults.tsx
@@ -1,10 +1,8 @@
 import { ArtistAuctionResults_artist } from "__generated__/ArtistAuctionResults_artist.graphql"
 import React, { Component } from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import { PaginationFragmentContainer as Pagination } from "Styleguide/Components/Pagination"
-import { Box } from "Styleguide/Elements/Box"
-import { Col, Row } from "Styleguide/Elements/Grid"
-import { Separator } from "Styleguide/Elements/Separator"
+import { PaginationFragmentContainer as Pagination } from "Styleguide/Components"
+import { Box, Col, Row, Separator } from "Styleguide/Elements"
 import { Subscribe } from "unstated"
 import { ArtistAuctionDetailsModal } from "./ArtistAuctionDetailsModal"
 import { AuctionResultItemFragmentContainer as AuctionResultsItem } from "./ArtistAuctionResultItem"
