+++ src/Apps/Artist/Routes/Overview/Components/ArtworkFilter/ArtworkFilterArtworkGrid.tsx
@@ -4,7 +4,7 @@ import { ContextConsumer } from "Artsy/Router"
 import ArtworkGrid from "Components/ArtworkGrid"
 import React, { Component } from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import { PaginationFragmentContainer as Pagination } from "Styleguide/Components/Pagination"
+import { PaginationFragmentContainer as Pagination } from "Styleguide/Components"
 import { Box } from "Styleguide/Elements/Box"
 import { Spacer } from "Styleguide/Elements/Spacer"
 import { Subscribe } from "unstated"
