+++ src/Apps/Artist/Routes/RelatedArtists/RelatedArtistsList.tsx
@@ -2,12 +2,13 @@ import { RelatedArtistsList_artist } from "__generated__/RelatedArtistsList_arti
 import { ContextConsumer } from "Artsy/Router"
 import React, { Component } from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
-import { ArtistCardFragmentContainer as ArtistCard } from "Styleguide/Components/ArtistCard"
-import { PaginationFragmentContainer as Pagination } from "Styleguide/Components/Pagination"
-import { Box } from "Styleguide/Elements/Box"
-import { Flex } from "Styleguide/Elements/Flex"
-import { Col, Row } from "Styleguide/Elements/Grid"
+import { Box, Col, Flex, Row } from "Styleguide/Elements"
 import { Responsive } from "Utils/Responsive"
+
+import {
+  ArtistCardFragmentContainer as ArtistCard,
+  PaginationFragmentContainer as Pagination,
+} from "Styleguide/Components"
 
 import {
   LoadingArea,
