+++ src/Apps/Artist/__static__/ShowsStaticLayout.tsx
@@ -4,12 +4,8 @@ import { showBlockProps, showListProps } from "Apps/__test__/Fixtures/Show"
 import { ArtistShowBlockItem } from "Apps/Artist/Routes/Shows/ArtistShowBlockItem"
 import { ArtistShowListItem } from "Apps/Artist/Routes/Shows/ArtistShowListItem"
 import React from "react"
-import { PaginationFragmentContainer } from "Styleguide/Components/Pagination"
-import { Box } from "Styleguide/Elements/Box"
-import { Flex } from "Styleguide/Elements/Flex"
-import { Col, Row } from "Styleguide/Elements/Grid"
-import { Separator } from "Styleguide/Elements/Separator"
-import { Spacer } from "Styleguide/Elements/Spacer"
+import { PaginationFragmentContainer } from "Styleguide/Components"
+import { Box, Col, Flex, Row, Separator, Spacer } from "Styleguide/Elements"
 import { Responsive } from "Utils/Responsive"
 
 const ShowBlocks = Flex
