+++ src/Apps/Artwork/ArtworkApp.tsx
@@ -5,9 +5,8 @@ import { artistResponse } from "Apps/__test__/Fixtures/MarketInsights"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components"
-import { FullArtworkGrid } from "Styleguide/Components/ArtworkGrid"
-import { Tab, Tabs } from "Styleguide/Components/Tabs"
-import { Col, Row } from "Styleguide/Elements/Grid"
+import { FullArtworkGrid, Tab, Tabs } from "Styleguide/Components"
+import { Col, Row } from "Styleguide/Elements"
 import { AboutTheWork } from "./Components/AboutTheWork"
 import { ArtistInfo } from "./Components/ArtistInfo"
 import { ArtworkSidebarFragmentContainer as ArtworkSidebar } from "./Components/ArtworkSidebar"
