+++ src/Styleguide/Components/Stepper.tsx
@@ -3,8 +3,8 @@ import { CheckIcon } from "Assets/Icons/CheckIcon"
 import { ChevronIcon } from "Assets/Icons/ChevronIcon"
 import React from "react"
 import styled from "styled-components"
-import { Tab, Tabs, TabsProps } from "Styleguide/Components/Tabs"
-import { Flex } from "Styleguide/Elements/Flex"
+import { Tab, Tabs, TabsProps } from "Styleguide/Components"
+import { Flex } from "Styleguide/Elements"
 import { styles } from "./Tabs"
 
 interface StepperProps extends TabsProps {
