+++ src/Styleguide/Components/__stories__/Lightbox.story.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
-import { Lightbox } from "Styleguide/Components/Lightbox"
-import { Image } from "Styleguide/Elements/Image"
+import { Lightbox } from "Styleguide/Components"
+import { Image } from "Styleguide/Elements"
 import { Section } from "Styleguide/Utils/Section"
 
 const image = size =>
