+++ src/Styleguide/Components/__stories__/RecentlyViewed.story.tsx
@@ -1,8 +1,8 @@
 import { artworkBricks } from "Apps/__test__/Fixtures/Slider"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
-import { RecentlyViewed } from "Styleguide/Components/RecentlyViewed"
-import { Box } from "Styleguide/Elements/Box"
+import { RecentlyViewed } from "Styleguide/Components"
+import { Box } from "Styleguide/Elements"
 import { Section } from "Styleguide/Utils/Section"
 
 storiesOf("Styleguide/Components", module).add("Recently Viewed", () => {
