+++ src/Styleguide/Components/__stories__/Tabs.Router.story.tsx
@@ -1,7 +1,7 @@
 import { StorybooksRouter } from "Artsy/Router"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
-import { RouteTab, RouteTabs } from "Styleguide/Components/RouteTabs"
+import { RouteTab, RouteTabs } from "Styleguide/Components"
 import { Section } from "Styleguide/Utils/Section"
 
 storiesOf("Styleguide/Components", module).add("Tabs (Router)", () => {
